+++ src/commands/general/start.ts
@@ -5,17 +5,16 @@ import {ISimulatorService} from "../../lib/interfaces/ISimulatorService";
 export interface StartActionOptions {
   resetValidators: boolean;
   numValidators: number;
-  branch: string;
-  location: string;
   headless: boolean;
+  resetDb: boolean
 }
 
 export async function startAction(options: StartActionOptions, simulatorService: ISimulatorService) {
-  const {resetValidators, numValidators, branch, location, headless} = options;
-  // Update simulator location with user input
-  simulatorService.setComposeOptions(headless);
-  simulatorService.setSimulatorLocation(location);
+  const {resetValidators, numValidators, headless, resetDb} = options;
 
+  simulatorService.setComposeOptions(headless);
+
+  await simulatorService.checkCliVersion();
 
   const restartValidatorsHintText = resetValidators
     ? `creating new ${numValidators} random validators`
@@ -23,17 +22,6 @@ export async function startAction(options: StartActionOptions, simulatorService:
 
   console.log(`Starting GenLayer simulator ${restartValidatorsHintText}`);
 
-  // Update the simulator to the latest version
-  console.log(`Updating GenLayer Simulator...`);
-  try {
-    await simulatorService.updateSimulator(branch);
-  } catch (error) {
-    console.error(error);
-    return;
-  }
-
-  // Run the GenLayer Simulator
-  console.log("Running the GenLayer Simulator...");
   try {
     await simulatorService.runSimulator();
   } catch (error) {
@@ -58,6 +46,10 @@ export async function startAction(options: StartActionOptions, simulatorService:
   } catch (error) {
     console.error(error);
     return;
+  }
+
+  if(resetDb){
+    await simulatorService.cleanDatabase()
   }
 
   if (resetValidators) {