+++ src/index.ts
@@ -1,11 +1,15 @@
 #!/usr/bin/env node
 import {program} from "commander";
 import {version} from "../package.json";
-import {CLI_DESCRIPTION} from "@/lib/config/text";
-import {initializeGeneralCommands} from "@/commands/general";
+import {CLI_DESCRIPTION} from "../src/lib/config/text";
+import { initializeGeneralCommands } from "../src/commands/general";
+import { initializeKeygenCommands } from "../src/commands/keygen";
 
-program.version(version).description(CLI_DESCRIPTION);
+export function initializeCLI() {
+  program.version(version).description(CLI_DESCRIPTION);
+  initializeGeneralCommands(program);
+  initializeKeygenCommands(program);
+  program.parse(process.argv);
+}
 
-initializeGeneralCommands(program);
-
-program.parse(process.argv);
+initializeCLI();